+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralium/FabricPeripheralium.kt
@@ -2,7 +2,10 @@ package site.siredvin.peripheralium
 
 import net.fabricmc.api.ModInitializer
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents
-import site.siredvin.peripheralium.fabric.*
+import site.siredvin.peripheralium.fabric.FabricIngredients
+import site.siredvin.peripheralium.fabric.FabricLibInnerPlatform
+import site.siredvin.peripheralium.fabric.FabricPeripheraliumPlatform
+import site.siredvin.peripheralium.fabric.FabricXplatTags
 import site.siredvin.peripheralium.storages.FabricStorageUtils
 import site.siredvin.peripheralium.storages.fluid.FluidStorageExtractor
 import site.siredvin.peripheralium.storages.item.ItemStorageExtractor